+++ codegen/projections/rails_json/lib/rails_json/errors.rb
@@ -14,18 +14,16 @@ module RailsJson
     end
 
     # Base class for all errors returned by this service
-    class ApiError < Hearth::HTTP::ApiError; end
+    class ApiError < Hearth::ApiError; end
 
     # Base class for all errors returned where the client is at fault.
-    # These are generally errors with 4XX HTTP status codes.
     class ApiClientError < ApiError; end
 
     # Base class for all errors returned where the server is at fault.
-    # These are generally errors with 5XX HTTP status codes.
     class ApiServerError < ApiError; end
 
     # Base class for all errors returned where the service returned
-    # a 3XX redirection.
+    # a redirection.
     class ApiRedirectError < ApiError
       def initialize(location:, **kwargs)
         @location = location
@@ -37,11 +35,11 @@ module RailsJson
     end
 
     class ComplexError < ApiClientError
-      def initialize(http_resp:, **kwargs)
-        @data = Parsers::ComplexError.parse(http_resp)
+      def initialize(data:, **kwargs)
+        @data = data
         kwargs[:message] = @data.message if @data.respond_to?(:message)
 
-        super(http_resp: http_resp, **kwargs)
+        super(**kwargs)
       end
 
       # @return [Types::ComplexError]
@@ -49,11 +47,11 @@ module RailsJson
     end
 
     class InvalidGreeting < ApiClientError
-      def initialize(http_resp:, **kwargs)
-        @data = Parsers::InvalidGreeting.parse(http_resp)
+      def initialize(data:, **kwargs)
+        @data = data
         kwargs[:message] = @data.message if @data.respond_to?(:message)
 
-        super(http_resp: http_resp, **kwargs)
+        super(**kwargs)
       end
 
       # @return [Types::InvalidGreeting]