+++ codegen/projections/white_label/lib/white_label/event_stream.rb
@@ -71,20 +71,6 @@ module WhiteLabel
         on(Types::Events::ExplicitPayloadEvent, block)
       end
 
-      # Register an event handler for server_error_event events
-      # @yield [event] Called when server_error_event events are received.
-      # @yieldparam event [Types::Events::ServerErrorEvent] the event.
-      # @example Event structure
-      #   event #=> Types::ServerErrorEvent
-      #   event.nested #=> Types::NestedStructure
-      #   event.nested.values #=> Array<String>
-      #   event.nested.values[0] #=> String
-      #   event.message #=> String
-      #   event.header_a #=> String
-      def on_server_error_event(&block)
-        on(Types::Events::ServerErrorEvent, block)
-      end
-
       # Register an event handler for any unknown events.
       # @yield [event] Called when unknown events are received.
       # @yieldparam event [Types::Events::Unknown] the event with value set to the Message
@@ -96,14 +82,37 @@ module WhiteLabel
 
       def parse_event(type, message)
         case type
-        when 'initial-response' then Parsers::EventStream::StartEventStreamInitialResponse.parse(message)
-        when 'SimpleEvent' then Types::Events::SimpleEvent.new(Parsers::EventStream::SimpleEvent.parse(message))
-        when 'NestedEvent' then Types::Events::NestedEvent.new(Parsers::EventStream::NestedEvent.parse(message))
-        when 'ExplicitPayloadEvent' then Types::Events::ExplicitPayloadEvent.new(Parsers::EventStream::ExplicitPayloadEvent.parse(message))
-        when 'ServerErrorEvent' then Types::Events::ServerErrorEvent.new(Parsers::EventStream::ServerErrorEvent.parse(message))
+        when 'initial-response'
+          Parsers::EventStream::StartEventStreamInitialResponse.parse(message)
+        when 'SimpleEvent'
+          Types::Events::SimpleEvent.new(Parsers::EventStream::SimpleEvent.parse(message))
+        when 'NestedEvent'
+          Types::Events::NestedEvent.new(Parsers::EventStream::NestedEvent.parse(message))
+        when 'ExplicitPayloadEvent'
+          Types::Events::ExplicitPayloadEvent.new(Parsers::EventStream::ExplicitPayloadEvent.parse(message))
+        when 'ServerErrorEvent'
+          Types::Events::ServerErrorEvent.new(Parsers::EventStream::ServerErrorEvent.parse(message))
         else
-          Types::Events::Unknown.new(name: type || 'unknown', value: message)
+          Types::Events::Unknown.new(name: type, value: message)
         end
+      end
+
+      def parse_exception_event(type, message)
+        case type
+        when 'ServerErrorEvent'
+          data = Parsers::EventStream::ServerErrorEvent.parse(message)
+          Errors::ServerErrorEvent.new(data: data, error_code: 'WhiteLabel::Types::Events::ServerErrorEvent')
+        else
+          data = Types::Events::Unknown.new(name: type, value: message)
+          Errors::ApiError.new(error_code: type, metadata: {data: data})
+        end
+      end
+
+      def parse_error_event(message)
+        error_code = message.headers.delete(':error-code')&.value
+        error_message = message.headers.delete(':error-message')&.value
+        metadata = {message: message}
+        Errors::ApiError.new(error_code: error_code, metadata: metadata, message: error_message)
       end
     end