+++ hearth/spec/hearth/event_stream/handler_base_spec.rb
@@ -30,6 +30,8 @@ module Hearth
       subject do
         Class.new(HandlerBase) do
           def parse_event(_, _); end
+          def parse_error_event(_); end
+          def parse_exception_event(_, _); end
         end.new
       end
 
@@ -66,13 +68,16 @@ module Hearth
           let(:message_type) { 'error' }
           let(:error_code) { 'error_code' }
           let(:error_message) { 'error_message' }
+          let(:error) { double }
 
           it 'calls registered error_event handlers' do
-            subject.on_error_event(&handler1)
-            subject.on_error_event(&handler2)
-            expect(handler1).to receive(:call).with(error_code, error_message)
-            expect(handler2).to receive(:call).with(error_code, error_message)
+            subject.on_error(&handler1)
+            subject.on_error(&handler2)
+            expect(handler1).to receive(:call).with(error)
+            expect(handler2).to receive(:call).with(error)
 
+            expect(subject).to receive(:parse_error_event)
+              .with(message).and_return(error)
             subject.emit(message)
           end
         end
@@ -83,9 +88,9 @@ module Hearth
           let(:exception_class) { Class.new }
           let(:exception_event) { exception_class.new }
 
-          it 'calls registered MyException handlers' do
-            subject.send(:on, exception_class, handler1)
-            expect(subject).to receive(:parse_event)
+          it 'calls registered error handlers' do
+            subject.on_error(&handler1)
+            expect(subject).to receive(:parse_exception_event)
               .with(exception_type, message)
               .and_return(exception_event)
             expect(handler1).to receive(:call).with(exception_event)